removed content from products
Removed entire 'Popular' drinks section from products page and also all h3 tags with 'See All Teas' written in them. That way I can make the whole products page just the small amount of drinks that I have, and also have a checkout button in the bottom of the page which will be displaying the total.